examples: Move helpers out of 'verify.js', in a separate file 'utils.js'
